requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
Jinja2==3.1.4
pandas==2.2.2
